Fix several issues with visual movements.
The fixes for Textadept 9.3 were not well tested and broke several
movements. Scintilla change 6111:d29f1dad85ac 'Display block caret
over the character at the end of a selection to be similar' made the
treatment of the block caret position inconsistent. Reverting this
resolved all of the movement issues. The goal of the Scintilla change
would be better resolved by treating the position consistently and
instead interpreting the selection as in the change.